Split scatter tests into static and dynamic tests.
The CUDA side handles only static shapes. To handle dynamic shapes the
ConvertToNVVM pass needs to handle shape dialect conversion to LLVM as
well as support the ABI for dynamic shapes. Here the support for
handling shape dialect is added by using the patterns that are used on
the CPU side.
For now, the scatter test is split into static and dynamic tests, with
dynamic tests not run on CUDA backend.
