+++ iree/compiler/Codegen/LLVMCPU/ConvertToLLVM.cpp
@@ -12,8 +12,6 @@
 #include "iree/compiler/Dialect/IREE/IR/IREEDialect.h"
 #include "iree/compiler/Dialect/IREE/IR/IREEOps.h"
 #include "iree/compiler/Dialect/Shape/IR/ShapeDialect.h"
-#include "iree/compiler/Dialect/Shape/IR/ShapeOps.h"
-#include "iree/compiler/Dialect/Shape/IR/ShapeTypes.h"
 #include "llvm/Support/raw_ostream.h"
 #include "mlir/Analysis/DataLayoutAnalysis.h"
 #include "mlir/Conversion/AffineToStandard/AffineToStandard.h"
@@ -540,76 +538,6 @@ class RemoveHALInterfaceOpPattern : public ConvertToLLVMPattern {
   }
 };
 
-class RemoveMakeRankedShape : public ConvertToLLVMPattern {
- public:
-  explicit RemoveMakeRankedShape(MLIRContext *context,
-                                 LLVMTypeConverter &typeconverter)
-      : ConvertToLLVMPattern(Shape::MakeRankedShapeOp::getOperationName(),
-                             context, typeconverter) {}
-
-  LogicalResult matchAndRewrite(
-      Operation *op, ArrayRef<Value> operands,
-      ConversionPatternRewriter &rewriter) const override {
-    // Check users are ops are going to be folded away.
-    for (auto user : op->getUsers()) {
-      if (!cast<Shape::TieShapeOp>(user) && !cast<Shape::RankedDimOp>(user))
-        return failure();
-    }
-    rewriter.eraseOp(op);
-    return success();
-  }
-};
-
-// Updates memref descriptors shape and strides informations and fold tie_shape
-// into updated memref descriptor.
-class ConvertTieShapePattern : public ConvertToLLVMPattern {
- public:
-  explicit ConvertTieShapePattern(MLIRContext *context,
-                                  LLVMTypeConverter &typeconverter)
-      : ConvertToLLVMPattern(Shape::TieShapeOp::getOperationName(), context,
-                             typeconverter) {}
-
-  LogicalResult matchAndRewrite(
-      Operation *op, ArrayRef<Value> operands,
-      ConversionPatternRewriter &rewriter) const override {
-    auto tieShapeOp = cast<Shape::TieShapeOp>(op);
-    auto loc = tieShapeOp.getLoc();
-    MemRefDescriptor sourceMemRef(operands.front());
-    auto makeRankedShapeOp =
-        cast<Shape::MakeRankedShapeOp>(tieShapeOp.shape().getDefiningOp());
-    auto rankedShapeType = makeRankedShapeOp.shape()
-                               .getType()
-                               .dyn_cast_or_null<Shape::RankedShapeType>();
-    if (!rankedShapeType) return failure();
-
-    auto shape = rankedShapeType.getAllDims();
-
-    // Update memref descriptor shape and strides.
-    // dynamic dim maybe in mid location, like shapex.ranked_shape<[128,?,128]
-    int dynIdx = 0;
-    for (int i = 0; i < shape.size(); ++i) {
-      if (shape[i] == ShapedType::kDynamicSize) {
-        sourceMemRef.setSize(rewriter, loc, i,
-                             makeRankedShapeOp.dynamic_dimensions()[dynIdx++]);
-      } else {
-        sourceMemRef.setConstantSize(rewriter, loc, i, shape[i]);
-      }
-    }
-    // Compute and update memref descriptor strides. Assumption here is memrefs
-    // are row-major e.g following index linearization x[i, j, k] = i * x.dim[1]
-    // * x.dim[2] + j * x.dim[2] + k
-    sourceMemRef.setConstantStride(rewriter, loc, shape.size() - 1, 1);
-    for (int i = shape.size() - 2; i >= 0; --i) {
-      auto stride = sourceMemRef.stride(rewriter, loc, i + 1);
-      auto dim = sourceMemRef.size(rewriter, loc, i + 1);
-      Value strideVal = rewriter.create<LLVM::MulOp>(loc, stride, dim);
-      sourceMemRef.setStride(rewriter, loc, i, strideVal);
-    }
-    rewriter.replaceOp(tieShapeOp, {sourceMemRef});
-    return success();
-  }
-};
-
 class ConvertToLLVMPass : public ConvertToLLVMBase<ConvertToLLVMPass> {
  public:
   ConvertToLLVMPass(std::string targetTriple = "",
@@ -684,9 +612,10 @@ void ConvertToLLVMPass::runOnOperation() {
   options.dataLayout = llvm::DataLayout(targetDataLayout);
   options.overrideIndexBitwidth(options.dataLayout.getPointerSizeInBits());
   LLVMTypeConverter converter(&getContext(), options, &dataLayoutAnalysis);
-  converter.addConversion([](Shape::RankedShapeType, SmallVectorImpl<Type> &) {
-    return success();
-  });
+  // converter.addConversion([](Shape::RankedShapeType, SmallVectorImpl<Type> &)
+  // {
+  //   return success();
+  // });
 
   OwningRewritePatternList patterns(&getContext());
 
@@ -710,6 +639,7 @@ void ConvertToLLVMPass::runOnOperation() {
   populateVectorToLLVMMatrixConversionPatterns(converter, patterns);
   populateVectorToLLVMConversionPatterns(converter, patterns);
   populateLinalgToLLVMConversionPatterns(converter, patterns);
+  populateShapeToLLVMConversionPatterns(&getContext(), &converter, patterns);
 
   // clang-format off
   patterns.insert<
@@ -719,9 +649,7 @@ void ConvertToLLVMPass::runOnOperation() {
     ConvertHALInterfaceWorkgroupCountOp,
     ConvertHALInterfaceLoadConstant,
     ConvertHALInterfaceBindingSubspanOp,
-    RemoveHALInterfaceOpPattern,
-    ConvertTieShapePattern,
-    RemoveMakeRankedShape
+    RemoveHALInterfaceOpPattern
   >(&getContext(), converter);
   // clang-format on
