+++ iree/test/e2e/xla_ops/BUILD
@@ -77,6 +77,7 @@ iree_check_single_backend_test_suite(
         include = ["*.mlir"],
         exclude = [
             "round.mlir",
+            "scatter_dynamic.mlir",
         ],
     ),
     compiler_flags = ["-iree-input-type=mhlo"],
@@ -136,6 +137,7 @@ iree_check_single_backend_test_suite(
             "reverse.mlir",
             "rsqrt.mlir",
             "scatter.mlir",
+            "scatter_dynamic.mlir",
             "select.mlir",
             "sine.mlir",
             "slice.mlir",
@@ -214,6 +216,7 @@ iree_check_single_backend_test_suite(
             "fft.mlir",  # TODO(GH-6601): Eanble the test.
             "round.mlir",
             "scatter.mlir",  # TODO(GH-6601): Enable the test.
+            "scatter_dynamic.mlir",  # TODO(GH-6601): Enable the test.
             "sort.mlir",
         ],
     ),
@@ -279,6 +282,7 @@ iree_check_single_backend_test_suite(
             "fft.mlir",  # TODO(GH-6388): Enable the test.
             "round.mlir",
             "scatter.mlir",  # TODO(GH-6388): Enable the test.
+            "scatter_dynamic.mlir",  # TODO(GH-6388): Enable the test.
             "sort.mlir",
         ],
     ),
