+++ iree/test/e2e/xla_ops/CMakeLists.txt
@@ -120,6 +120,7 @@ iree_check_single_backend_test_suite(
     "reverse.mlir"
     "rsqrt.mlir"
     "scatter.mlir"
+    "scatter_dynamic.mlir"
     "select.mlir"
     "sine.mlir"
     "slice.mlir"
